Refactor the server validations package
- better names
- simpler implementation
