+++ pkg/server/validation/email.go
@@ -1,0 +1,8 @@
+package validation
+
+import "net/mail"
+
+func IsValidEmail(email string) bool {
+	_, err := mail.ParseAddress(email)
+	return err == nil
+}
